requirements.txt
aiosqlite==0.19.0
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
click==8.1.7
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
passlib==1.7.4
pluggy==1.4.0
pydantic==2.6.0
pydantic_core==2.16.1
pytest==7.4.4
pytest-asyncio==0.23.4
python-magic==0.4.27
python-multipart==0.0.6
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.27.0.post1
